Move prepare into analysis/main.js.
